% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder.R
\name{add_json_format}
\alias{add_json_format}
\title{Adds the json formatting option to the
passed url request.}
\usage{
add_json_format(url)
}
\arguments{
\item{url}{request url to which the json format option
should be added}
}
\value{
url with json format option added
}
\description{
Adds the json formatting option to the
passed url request.
}
