% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder.R
\name{add_path_part}
\alias{add_path_part}
\title{Adds a path part to the given url}
\usage{
add_path_part(url, path_name)
}
\arguments{
\item{url}{url to which a path part should
be added}

\item{path_name}{path part which should be added
to the url}
}
\value{
url with the path part added
}
\description{
Adds a path part to the given url
}
