% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder.R
\name{create_request}
\alias{create_request}
\title{Creates a request with the given path parts.
The json format argument is included by default.}
\usage{
create_request(base_url, path_parts = NULL)
}
\arguments{
\item{base_url}{base url of the API for which a request
should be created}

\item{path_parts}{that should be added to the base url}
}
\value{
request url composed of the given base url and
specified path parts
}
\description{
Creates a request with the given path parts.
The json format argument is included by default.
}
