% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_goldprice.R
\name{.goldprice_base_url}
\alias{.goldprice_base_url}
\title{Returns the base url for the gold price endpoint.}
\usage{
.goldprice_base_url()
}
\description{
Returns the base url for the gold price endpoint.
}
