% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_rates.R
\name{.send_rates_endpoint_request}
\alias{.send_rates_endpoint_request}
\title{Sends a request and parses the rates endpoint response.}
\usage{
.send_rates_endpoint_request(request_url)
}
\arguments{
\item{request_url}{url to which the request should be sent.}
}
\value{
nbp_api_response object with the request content.
}
\description{
Sends a request and parses the rates endpoint response.
}
