% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint_goldprice.R
\name{get_current_goldprice}
\alias{get_current_goldprice}
\title{Retrieves the current gold price.}
\usage{
get_current_goldprice()
}
\value{
nbp_api_response object containing the current
gold price.
}
\description{
Retrieves the current gold price.
}
\examples{

\donttest{
## Fetch the current gold price
response <- get_current_goldprice()

## Retrieve the current gold price value
response$content$cena

}

}
\seealso{
\url{https://api.nbp.pl/#cenyZlota}

Other goldprice: 
\code{\link{get_goldprice_from_interval}()},
\code{\link{get_goldprice_from}()},
\code{\link{get_last_n_goldprices}()},
\code{\link{get_todays_goldprice}()}
}
\concept{goldprice}
