% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testers.R
\name{is_count}
\alias{is_count}
\title{Checks if an object is a positive integer.}
\usage{
is_count(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
TRUE or FALSE depending on whether its
argument is a positive integer or not.
}
\description{
Checks if an object is a positive integer.
}
