% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testers.R
\name{is_date}
\alias{is_date}
\title{Checks if an object is a date object.}
\usage{
is_date(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
TRUE or FALSE depending on whether its
arguments is a date object or not.
}
\description{
Checks if an object is a date object.
}
