% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_testers.R
\name{is_integer}
\alias{is_integer}
\title{Checks if an object is an integer.}
\usage{
is_integer(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
TRUE or FALSE depending on whether its
argument is an integer or not.
}
\description{
Checks if an object is an integer.
}
